requirements.txt
cartopy==0.22.0
matplotlib==3.8.2
mayavi==4.8.1
networkx==3.1
numpy==1.25.2
pandas==1.5.3
pandasai==1.5.19
Pillow==10.2.0
python-dotenv==1.0.1
scipy==1.12.0
seaborn==0.13.2
streamlit==1.30.0
openpyxl==3.1.2